Move TrackArtists field to separate component
