+++ src/components/TrackFormArtists.vue
@@ -1,0 +1,134 @@
+<template>
+  <div>
+    <VRow :key="index" v-for="(item, index) of trackArtists" no-gutters>
+      <VCol class="flex-column flex-grow-0">
+        <VBtn
+          @click="moveArtist(index, -1)"
+          icon
+          small
+          class="ma-2"
+          :disabled="index === 0"
+        >
+          <VIcon>mdi-menu-up</VIcon>
+        </VBtn>
+        <VBtn
+          @click="moveArtist(index, 1)"
+          icon
+          small
+          class="ma-2"
+          :disabled="index === trackArtists.length - 1"
+        >
+          <VIcon>mdi-menu-down</VIcon>
+        </VBtn>
+        <VBtn @click="removeArtist(index)" icon small class="ma-2">
+          <VIcon>mdi-close</VIcon>
+        </VBtn>
+      </VCol>
+      <VCol>
+        <VCombobox
+          :items="sortedArtists"
+          :filter="filterName"
+          item-text="name"
+          item-value="id"
+          :label="$tc('music.artists', 2)"
+          :rules="rules"
+          return-object
+          v-model="item.artist_id"
+        />
+        <VTextField :label="$t('common.name')" v-model="item.name" />
+        <VAutocomplete
+          :items="roles"
+          :label="$t('music.artist.role')"
+          v-model="item.role"
+        />
+        <VDivider light v-if="index !== trackArtists.length - 1" />
+      </VCol>
+    </VRow>
+  </div>
+</template>
+
+<script>
+import { mapGetters } from "vuex";
+
+export default {
+  name: "TrackFormArtists",
+  data() {
+    return {
+      roles: [
+        {
+          value: "main",
+          text: this.$t("music.artist.roles.main"),
+        },
+        {
+          value: "performer",
+          text: this.$t("music.artist.roles.performer"),
+        },
+        {
+          value: "composer",
+          text: this.$t("music.artist.roles.composer"),
+        },
+        {
+          value: "conductor",
+          text: this.$t("music.artist.roles.conductor"),
+        },
+        {
+          value: "remixer",
+          text: this.$t("music.artist.roles.remixer"),
+        },
+        {
+          value: "producer",
+          text: this.$t("music.artist.roles.producer"),
+        },
+        {
+          value: "arranger",
+          text: this.$t("music.artist.roles.arranger"),
+        },
+      ],
+      trackArtists: [],
+    };
+  },
+  props: {
+    value: {
+      type: Array,
+      required: true,
+    },
+  },
+  mounted() {
+    this.trackArtists = this.value;
+  },
+  watch: {
+    trackArtists(newValue) {
+      this.$emit("update:value", newValue);
+    },
+  },
+  computed: {
+    ...mapGetters("artists", {
+      sortedArtists: "artistsByName",
+    }),
+    rules() {
+      const artistValidation = (v) =>
+        !!v || this.$t("errors.artists.artist-blank");
+      return [artistValidation];
+    },
+  },
+  methods: {
+    filterName(item, queryText) {
+      const search = queryText.toLowerCase();
+      return (
+        item.name.toLowerCase().indexOf(search) > -1 ||
+        item.normalized_name.indexOf(search) > -1
+      );
+    },
+    removeArtist(index) {
+      this.trackArtists.splice(index, 1);
+    },
+    moveArtist(index, direction) {
+      this.trackArtists.splice(
+        index + direction,
+        0,
+        this.trackArtists.splice(index, 1)[0]
+      );
+    },
+  },
+};
+</script>
